functions/requirements.txt
ydb~=2.3.1
fire~=0.4.0
yandexcloud
bcrypt~=3.2.0
requests~=2.27.1
PyJWT==2.3.0
json-rpc-stateful~=1.13.3